isolate EditTagModal
